// wetty-chat-mobile/src/generated/models/messageResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wetty Chat API
 * Real-time chat application backend API supporting groups, messaging, threads, stickers, invites, and push notifications.
 * OpenAPI spec version: 0.1.0
 */
import type { AttachmentResponse } from './attachmentResponse';
import type { MentionInfo } from './mentionInfo';
import type { MessagePreview } from './messagePreview';
import type { MessageStickerResponse } from './messageStickerResponse';
import type { MessageType } from './messageType';
import type { ReactionSummary } from './reactionSummary';
import type { SnowflakeID } from './snowflakeID';
import type { ThreadInfo } from './threadInfo';
import type { User } from './user';

export interface MessageResponse {
  attachments: AttachmentResponse[];
  chatId: SnowflakeID;
  clientGeneratedId: string;
  createdAt: string;
  hasAttachments: boolean;
  id: SnowflakeID;
  isDeleted: boolean;
  isEdited: boolean;
  mentions?: MentionInfo[];
  message?: string;
  messageType: MessageType;
  reactions: ReactionSummary[];
  replyRootId?: SnowflakeID;
  replyToMessage?: MessagePreview;
  sender: User;
  sticker?: MessageStickerResponse;
  threadInfo?: ThreadInfo;
}